+++ Debug/Debug/Src/main.cyclo
@@ -1,0 +1,4 @@
+../Src/main.c:12:6:finished_transmission	2
+../Src/main.c:23:6:USART1_IRQHandler	2
+../Src/main.c:34:6:message_received	1
+../Src/main.c:42:5:main	2
