+++ Debug/Debug/Src/main.su
@@ -1,0 +1,4 @@
+../Src/main.c:12:6:finished_transmission	24	static
+../Src/main.c:23:6:USART1_IRQHandler	16	static
+../Src/main.c:34:6:message_received	16	static
+../Src/main.c:42:5:main	40	static
